% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bing_recording_data.R
\name{get_bing_recording_data}
\alias{get_bing_recording_data}
\title{get_bing_recording_data}
\usage{
get_bing_recording_data()
}
\arguments{
\item{include_luna}{Set to \code{TRUE} to include Luna data.}
}
\value{
Data frame of bv-main recordings.
}
\description{
Gets all the data in bv-bing, which is the Bing headcam study.
}
