% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_main_recording_data.R
\name{get_main_recording_data}
\alias{get_main_recording_data}
\title{get_main_recording_data}
\usage{
get_main_recording_data(include_luna = FALSE)
}
\arguments{
\item{include_luna}{Set to \code{TRUE} to include Luna data.}
}
\value{
Data frame of bv-main recordings.
}
\description{
Gets all the data in bv-main, which is the home headcam study
}
