% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participant_data.R
\name{get_participant_data}
\alias{get_participant_data}
\title{get_participant_data}
\usage{
get_participant_data(include_demographics = TRUE)
}
\arguments{
\item{include_demogrpahics}{Set to \code{TRUE} to merge demographic information.}
}
\value{
Data frame of participants.
}
\description{
Gets all the data in the participant table, optionally merging demographics.
}
